refactor: generator 패키지 생성
