+++ src/main/java/ladder/LadderConsoleApp.java
@@ -1,10 +1,6 @@
 package ladder;
 
 import ladder.controller.LadderGameController;
-import ladder.domain.GamePlayers;
-import ladder.domain.Ladder;
-import ladder.domain.PlayerGenerator;
-import ladder.view.OutputConsoleView;
 
 public class LadderConsoleApp {
     public static void main(String[] args) {
