+++ src/main/java/ladder/domain/generator/LineGenerator.java
@@ -1,4 +1,6 @@
-package ladder.domain;
+package ladder.domain.generator;
+
+import ladder.domain.Line;
 
 import java.util.List;
