+++ src/main/java/ladder/domain/generator/LineRandomGenerator.java
@@ -1,4 +1,6 @@
-package ladder.domain;
+package ladder.domain.generator;
+
+import ladder.domain.Line;
 
 import java.util.ArrayList;
 import java.util.List;
