+++ src/main/java/ladder/domain/generator/SubLineRandomGenerator.java
@@ -1,6 +1,5 @@
-package ladder.domain;
+package ladder.domain.generator;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Random;
 import java.util.Stack;
